refactor: シェイク強度計算をmapEffectsに移動
calcScreenShakeIntensityと関連定数をpopupLogic.tsから
mapEffects.tsに移動し、テストもmapEffects.test.tsに移動。

Closes #515
